+++ docs/qubinode/releases.md
@@ -5,6 +5,27 @@
 ### Release web page
 https://github.com/Qubinode/qubinode-installer/releases
 
+**Qubinode v2.4.3 Release notes**
+* OpenShift 4.6.x support
+* fix IDM Bugs
+* Local Storage Option Recommend deployment option
+* External Bridge Deployment Support for OpenShift Nodes
+
+**Qubinode v2.4.2 Release notes**
+* fixed NFS server bug adding nfs server mount to the correct location `/home/nfs_mount/data`
+* adding NFS Server as default for remote storage
+
+
+**Qubinode v2.4.2 Release notes**
+* RHEL 8.2 Support
+* OpenShift sizing menu fixes
+* OpenShift 4.4.x
+* OpenShift 4.5.x (Optional)
+* OKD 4 menu Option
+* Ansible 2.9 Compatibility
+* local storage support mutliple disk support
+* Fixed IDM bugs
+* Ansible Tower Support
 
 **Qubinode v2.4.2 Release notes**
 * RHEL 8.2 Support